# fix RowwiseMoments vectorization issue on CPU