@malfet 2022-08-11
@pytorchbot revert -c weird -m "Regressed performance, see benchmark issue"